requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-search-documents==11.4.0
azure-ai-vision-imageanalysis==1.0.0b1
azure-core==1.29.5
azure-storage-blob==12.19.0
openai==1.3.7
redis[hiredis]==5.0.1
python-multipart==0.0.6
Pillow==10.1.0